Reformatted code under new coding style, some braces next line now. Forced aligned vars. Moved messaging into interaction messager class Created progress class for help doing this like progress bars Added progress logging to piratebaylocal Added description of commands above them Added new torrent source exception for special catching Made string percent based matching not case sensitive Added view system (pagination) to torrents so now they are not all displayed but are all loaded and associated commands. Converted TorrentLoaded to work as instance based and handle pagination Fixed a bug that I've now completely forgotten about.

+++ Server/src/com/mt/easytv/torrents/Torrent.java
@@ -4,14 +4,13 @@ public class Torrent
 {
     public  String          url;
     public  String          name;
-    public int size; //bytes
-    public int seeders;
+    public float size; //bytes
+    public int   seeders;
     public  int             leechers;
     public  int             age;
     private TorrentDownload _download;
 
-    public TorrentDownload getDownload()
-    {
+    public TorrentDownload getDownload() {
         if (this._download == null) {
             this._download = new TorrentDownload(this.url);
         }
@@ -19,8 +18,7 @@ public class Torrent
         return this._download;
     }
 
-    public String toString()
-    {
+    public String toString() {
         return this.name + " (" + this.url + ") size of " + this.size + " with " + this.seeders + "/" + this.leechers + "(S/L) ratio.";
     }
 }
